Rename intervention.py to user_action.py and update imports
